complete bicycle model with noise
move applies bicycle model to move particle depending on d, distance to move forward, and alpha, steering wheel angle. Resulting state is calculated with following noise:
for move:
   alpha=gaussian(motion[0],steering_noise)
   d=motion[1]+gaussian(0,distance_noise)

sensing gives the bearing to each landmark. in this world, landmarks are four corners of square world. Noise is added to the global robot orientation thus:
for sense:
   orientation+=gaussian(0,bearing_noise)